+++ source2/Kernel/System/Cosmos.System/Filesystem/Listing/File.cs
@@ -5,9 +5,8 @@ using System.Text;
 
 namespace Cosmos.System.Filesystem.Listing {
   public class File : Base {
-    //TODO:UInt32? Size until ulong works.. nullable types have a problem right now
-    //TODO:UInt64? Size
-    public File(FileSystem aFileSystem, string aName, UInt32 aSize)
+    //TODO:UInt64? Size until ulong works.. nullable types have a problem right now
+	public File(FileSystem aFileSystem, string aName, UInt64 aSize)
       : base(aFileSystem, aName) {
       mSize = aSize;
     }
